restructure tag command
